+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 4.3.0
+* 중요 
+* 간편결제는 생체인증, 통합결제 2가지를 사용중 
+* 간편결제시 서버 분리 승인 로직을 추가하면서 사이드 이펙트가 생긴것을 확인하여 아래와 같이 수정함
+  * 발견된 사이드 이펙트는 서버 승인 로직이 추가되면서 클라이언트에서 간편결제 승인이 진행되지 않는 것임 (separatelyConfirmed 옵션 기본값이 true)
+  * 수정사항) extra -> bio_extra 모델로 대체 
+    * bio_extra.separatelyConfirmed 옵션은 통합결제에만 적용되도록 수정 
+    * bio_extra.separatelyConfirmedBio 옵션은 간편결제에만 적용되도록 수정 
+      - 단 이 옵션이 true일 경우 특성상 서버승인으로만 진행을 해야함 
+
 ## 4.2.9
 * bootpay js 4.2.2 적용 
 * bootpay 4.2.7 적용